% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{annotation_plot}
\alias{annotation_plot}
\title{annotation plots for CpG sites}
\usage{
annotation_plot(
  result,
  id,
  phenotype,
  width = 50000,
  FDR = F,
  verbose = T,
  print_log = F,
  log_path = "log.txt",
  save_dest = "C:/Users/weihsa/Documents/"
)
}
\arguments{
\item{result}{(data.frame) data set created by \code{\link{meta}}}

\item{id}{(string) CpG ID of the CpG site that should be plotted}

\item{phenotype}{(string) name of the phenotype analysed}

\item{width}{(int) width of the window size around the CpG site that should be plotted}

\item{FDR}{(bool) TRUE: uses FDR correction FALSE: won't}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log file}

\item{save_dest}{(string) path to where the plots should be saved}
}
\description{
plotting function that creates annotation plots of CpG sites using the UCSC database
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
