% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rma_uni.R
\name{auto_rma_uni}
\alias{auto_rma_uni}
\title{auto_rma_uni()}
\usage{
auto_rma_uni(data, group = FALSE, effect = "DL")
}
\arguments{
\item{data}{Object of class "escalc" from metafor package.}

\item{group}{Group column, if data should be grouped. Default is FALSE.}

\item{effect}{Model to calculate the effect according
to metafor::rma.uni. Default is "DL".}
}
\value{
A list containing objects of class rma.uni
or a unique rma.uni object.
}
\description{
List of rma.uni objects created according to subgroups.
}
\examples{
library(metafor)
rma_uni_list <- auto_rma_uni(dat.riley2003, group = "outcome", effect = "DL")
}
