% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_bayes.R
\name{bayes_post_process}
\alias{bayes_post_process}
\title{Bayes meta-analysis post process}
\usage{
bayes_post_process(
  result,
  combined_data,
  run_posterior_Check = TRUE,
  cutoff = 0.95,
  replicates = 1000,
  plot_forest = TRUE,
  return_hits_only = FALSE,
  output_path = "./",
  annotate_result = TRUE,
  phenotype = NULL,
  stratum = NULL,
  print_log = FALSE,
  log_path = "./",
  verbose = TRUE,
  num_cores = NULL
)
}
\arguments{
\item{result}{(data.frame) result from \code{\link{meta}} using the Bayesian model}

\item{combined_data}{(data.frame) data used to create the above result from \code{\link{load_files}}}

\item{run_posterior_Check}{(bool) TRUE: will run a posterior predictive (see \link[bayesmeta]{pppvalue}) (NOTE: computationally intensive) FALSE: won't}

\item{cutoff}{(double) effect posterior probability cut-off}

\item{replicates}{(int) number of replications for the posterior probability check (n >> 100 recommended)}

\item{plot_forest}{(bool) TRUE: Draws a forest plot of the top hits FALSE: won't}

\item{return_hits_only}{(bool) TRUE: will only return sites above the significance level cut off. FALSE: will return the whole data set.}

\item{output_path}{(string) output path for the forest plot}

\item{annotate_result}{(bool) TRUE: will merge output file with annoation file FALSE: won't}

\item{phenotype}{(string) phenotype used in the analysis}

\item{stratum}{(string) stratum used in the analysis}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: Prints output to terminal FALSE: won't}

\item{num_cores}{(int) number of parallel runs (when \code{NULL}, the number of cores will be calculated automatically)}
}
\value{
\code{result} data frame containing all analysed sites (see \code{\link{meta}}) with the following extra columns
        \itemize{
        \item{\code{pp.applied}}{: TRUE if posterior check has been applied, FALSE else}
        \item{\code{pp.pValue}}{: p-value posterior predictive check (NA if effect posterior probability below \code{cutoff}) }
        \item{\code{pp.replicates}}{: number of replicates performed (equals the \code{replicates} variable)}
        \item{\code{pp.computation_time}}{: time required for the computation}
        }
}
\description{
post processing function, that, if(\code{plot_forest} is TRUE plots forest plots of the sites whose effect posterior is above \code{cutoff} and,
             if \code{run_posterior_check} is TRUE, runs a posterior predictive check on the results. NOTE: Beware of significant runtime
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
