% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{clean_data()}
\usage{
clean_data(data, clean_column = FALSE, case = "snake", allow_dupes = FALSE)
}
\arguments{
\item{data}{Dataframe containing the data.}

\item{clean_column}{Column to be cleaned according to the
"case" parameter. Default is FALSE.}

\item{case}{Case type from janitor::make_clean_names().
Dafault is "snake".}

\item{allow_dupes}{Allow duplicated names. Default is FALSE.}
}
\value{
Returns a cleaned dataset.
}
\description{
Clean the dataframe before the analysis.
}
\examples{
df <- readr::read_csv(file.choose())
df_cleaned <- clean_data(df)
}
