% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{double_manhattan}
\alias{double_manhattan}
\title{double Manhattan Plot}
\usage{
double_manhattan(
  x,
  chr = "CHR",
  bp = "BP",
  p = "P",
  markername = "MARKERNAME",
  beta = "BETA",
  col = c("gray10", "gray60"),
  FDRcorr = TRUE,
  cutoff = 0.05,
  strict_cutoff = 0.001,
  title = "Manhattan plot",
  logp = TRUE,
  marktop = FALSE,
  save_plot = TRUE,
  save_path = "./",
  ...
)
}
\arguments{
\item{x}{(data.frame) input data set}

\item{chr}{(string) name of the coloumn in \code{x} that contains the chromosome numbers}

\item{bp}{(string) name of the coloumn in \code{x} that contains the genome position}

\item{p}{(string) name of the coloumn in \code{x} that contains the p-values}

\item{markername}{(string) name of the column in \code{x} that contains the cpg site names}

\item{col}{(string vector) name of the two colours the manhattan plot shall be coloured in}

\item{FDRcorr}{(bool) TRUE: will perform an FDR correction on the p-values and use these FALSE: will use the p-values}

\item{cutoff}{(double) significance level}

\item{strict_cutoff}{(double) strict significance level}

\item{title}{(string) title of the manhattan plot}

\item{logp}{(bool) TRUE will use a logarithmic scale on the y-axis FALSE: won't}

\item{marktop}{(bool) TRUE: will label the cpg sites above the cutoff FALSE: won't}

\item{save_plot}{(bool) TRUE: saves the plot to \code{save_path} FALSE: won't}

\item{save_path}{(string) save path for the manhattan plot}
}
\description{
This function creates a double manahattan plot for \code{\link{post_process}}
             The base code was extracted from the manhattan() function from the qqman r package
}
\references{
Stephen D. Turner, 2018, "qqman: an R package for visualizing GWAS results usinf Q-Q and manhattan plots", Jorunal of Open Source Software, 3(25), 731, doi:10:21105/joss.00731
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
