% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files}
\alias{load_files}
\title{Load files for this package}
\usage{
load_files(data_summary_path, phenotype, stratum, cohort = NULL,
  FDR = TRUE, annotation = FALSE, anno_file_path = "./",
  verbose = TRUE, print_log = TRUE, log_path = "./log.txt")
}
\arguments{
\item{data_summary_path}{(string) path to the file that contains all the info of the cohorts. Has to include the following tab separated
columns: \itemize{
\item{cohort}{: Name of the cohort}
\item{phenotype}{: Name of the observed phenotype used in the file}
\item{stratum}{: Name of the stratum used in the file}
\item{array_type}{: type of array used by the cohort. Can be 450 or EPIC}
\item{probeID}{: Name of the column containing the probeIDs}
\item{beta}{: Name of the column containing the effect sizes}
\item{se}{: Name of the column containing the standard error values}
\item{p_val}{: Name of the column containing the p-values}
\item{size}{: Name of the column containing the sample sizes}
\item{separator}{: type of separator used in the file. Example: ","}
\item{filename}{: Full name of the file. Example: "example_name.csv"}
\item{file_path}{: Full path to the file. Example: "/home/user/file_location/"}
}}

\item{phenotype}{(string) phenotype that should be analysed (has to appear in the summary$phenotype column)}

\item{stratum}{(string) stratum that should be analysed (has to appear in the summary$stratum column)}

\item{cohort}{(string) if only certain cohorts of the summary file should be analysed (has to appear in the summary$cohort column)}

\item{FDR}{(bool) TRUE: Add a column to the output with Bonferroni corrected p-values FALSE: won't}

\item{annotation}{(bool) TRUE: Adds chromosome and MAPINFO column to output FALSE: won't}

\item{anno_file_path}{(string) path to annotation file (required if annotation is TRUE). NOTE:
- Has to include the following tab separated columns:
  IlmnID (contains the CpG site IDs), CHR (contains the chromosome number), MAPINFO (contains the MAPINFO)}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log}
}
\value{
data frame used by the rest of the package containing the combined data of the loaded cohorts
}
\description{
This function imports the required files and modifies them in such a way, that they can be used by the package.
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
