% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Meta-analysis function}
\usage{
meta(
  input_data,
  test_run = FALSE,
  test_run_length = 1000,
  num_cores = 1,
  model = "FE",
  cohort = NULL,
  cpg_filter = NULL,
  FDR = TRUE,
  print_log = FALSE,
  log_path = "log.txt",
  save_result = FALSE,
  save_dest = "./",
  save_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{input_data}{(data.frame) data set created by \code{\link{load_files}}}

\item{test_run}{(bool) FOR DEBUG TRUE: only the first \code{test_run_length} sites are analysed FALSE: everything is analysed}

\item{test_run_length}{FOR DEBUG (int) number of site to be run when \code{test_run} is TRUE}

\item{num_cores}{(int) number of cores that should be used (only works in Linux systems)}

\item{model}{(string) Estimator model for tau. Standard \code{FE}. Otherwise \code{BAYES} for a Bayesian approach, \code{REML} for a
restricted maximum likelihood approach (not recommended for small number of cohorts) or \code{Z} for a p-value based approach}

\item{cohort}{(vector string) order in which the cohorts should appear in the output. If null random order is used.}

\item{cpg_filter}{(vector string) a vector containing the CpG sites that should be analysed}

\item{FDR}{(bool) TRUE: p values will be corrected false discovery rate, FALSE: won't WARNING: If TRUE, only sites no errors will be returned}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log file}

\item{save_result}{(bool) TRUE: saves the result as *.RData FALSE: won't}

\item{save_dest}{(string) path to save spot}

\item{save_name}{(string) file name for the saved result}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}
}
\value{
data.frame containing the results of the meta-analysis. If the model is FE or REML the output will have the following columns:
        \itemize{
        \item{\code{Markername}}{: CpG Site ID}
        \item{\code{Model}}{: Name of the model used (either FE or REML)}
        \item{\code{Estimate_phenotype}}{: estimated effect}
        \item{\code{SE_phenotype}}{: corresponding standard error}
        \item{\code{Pval_phenotype}}{: corresponding p-value}
        \item{\code{Zval_phenotype}}{: corresponding z-value}
        \item{\code{Heterogeneity}}{: estimated heterogeneity (NA if model is FE)}
        \item{\code{I2}}{: I squared parameter. Describes the percentage of variation across studies that is due to heterogeneity rather than by chance}
        \item{\code{number_of_cohorts}}{: number of cohorts available for the site}
        \item{\code{Cohort Names}}{: Multiple columns (one for each cohort) with + if the cohort effect is positive, - if it is negative, 0 if the effect is zero and ? if the no effect is present}
        \item{\code{sample_size}}{: total number of samples included in this site}
        \item{\code{analysis_error}}{: TRUE if a problem occurred, FALSE if none occurred}
        \item{\code{error_message}}{: contains error message if analysis_error is TRUE}
        \item{\code{FDR}}{: exists if FDR is TRUE and contains the FDR adjusted p-values}
        }
        If the model is BAYES the output will have the following columns
        \itemize{
        \item{\code{Markername}}{: CpG Site ID}
        \item{\code{Model}}{: Name of the model used (in this case BAYES)}
        \item{\code{MAP_effect}}{: maximum a-posteriori effect estimate}
        \item{\code{MAP_tau}}{: maximum a-posteriori heterogeneity estimate}
        \item{\code{bayes_factor_tau_eq_0}}{: Bayes factor for the H0 hypothesis tau = 0}
        \item{\code{bayes_factor_effect_eq_0}}{: Bayes factor for the H0 hypothesis effect = 0}
        \item{\code{minimum_bayes_factor_tau_eq_0}}{: minimum Bayes factor for the H0 hypothesis tau = 0}
        \item{\code{minimum_bayes_factor_effect_eq_0}}{: Bayes factor for the H0 hypothesis effect = 0}
        \item{\code{post_probability_tau_smaler_than_one}}{: posterior probability of the heterogeneity being smaller than 1, at which stage the heterogeneity can be considered to be extreme}
        \item{\code{post_probability_tau_bigger_than_one}}{: posterior probability of the heterogeneity being bigger than 1, at which stage the heterogeneity can be considered to be extreme}
        \item{\code{post_probability_effect_smaler_than_zero}}{: posterior probability of the effect being smaller than zero}
        \item{\code{post_probability_effect_bigger_than_zero}}{: posterior probability of the effect being bigger than zero}
        \item{\code{time_taken}}{: analysis time take for this site in seconds}
        \item{\code{number_of_cohorts}}{: number of cohorts available for the site}
        \item{\code{sample_size}}{: total number of samples included in this site}
        \item{\code{Cohort Names}}{: Multiple columns (one for each cohort) with + if the cohort effect is positive, - if it is negative, 0 if the effect is zero and ? if the no effect is present}
        \item{\code{analysis_error}}{: TRUE if a problem occurred, FALSE if none occurred}
        \item{\code{error_message}}{: contains error message if analysis_error is TRUE}
        }
}
\description{
Function that performs the actual meta-analysis
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
