% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining_diamonds.R
\name{mining_diamonds}
\alias{mining_diamonds}
\title{mining_diamonds()}
\usage{
mining_diamonds(
  rma_summary,
  diamond_height = 0.1,
  subgroup = 1,
  main_group,
  second_level_grouping = FALSE,
  order
)
}
\arguments{
\item{rma_summary}{A df from rma_list_to_summary_df().}

\item{diamond_height}{Height of the diamond. Default is 0.1.}

\item{subgroup}{Number of subgroups. Default is 1; maximum is 2.}

\item{main_group}{String of the column name
that indicates the subgroup.}

\item{second_level_grouping}{Column name of the second level
group in case of two groups. Default is FALSE.}

\item{order}{Order of the group factors (ex. c("a", "b", "c").
The same order must be provided in the function
metaPipeline::show_me_diamonds().}
}
\value{
A dataframe containing diamond coordinates for the
ggplot2::geom_polygon().
}
\description{
Calculates diamond coordinates for the
ggplot2::geom_polygon().
}
\examples{
library(metafor)
order <- c("DFS", "OS")
diamonds <- metaPipeline::auto_rma_uni(dat.riley2003, group = "outcome") |>
   metaPipeline::rma_list_to_summary_df() |>
   metaPipeline::mining_diamonds(main_group = "outcome", order = order)
}
