% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diamonds.R
\name{plot_diamonds}
\alias{plot_diamonds}
\title{plot_diamonds()}
\usage{
plot_diamonds(
  coordinates_df,
  group1,
  group2 = "none",
  vline_ref = "none",
  title = "none",
  xlab = "none",
  show_grids_v = FALSE,
  show_grids_h = FALSE,
  ci_interval = FALSE,
  n_col = 3,
  adj1,
  adj2,
  order,
  ...
)
}
\arguments{
\item{coordinates_df}{A dataframe containing diamonds coordinates obtained
from metaPipeline::mining_diamonds().}

\item{group1}{String containing the variable used in y axis.}

\item{group2}{String containing the variable used in facets in case of
subgrouping. Default is "none".}

\item{vline_ref}{X axis position to plot the reference vertical line.
Default is "none".}

\item{title}{String containing the title of the plot. Default is "none".}

\item{xlab}{String containing the x axis title of the plot. Default is
"none".}

\item{show_grids_v}{Show vertical grids if TRUE. Default is FALSE.}

\item{show_grids_h}{Show horizontal grids if TRUE. Default is FALSE.}

\item{ci_interval}{If TRUE, plot data showing estimate, lower and upper
bounds and p-values. Default is FALSE.}

\item{n_col}{Number of columns in case group 2 is used. If ci_interval is
used with group2, a single column should be used.}

\item{adj1}{Value controlling the distance between the graph area and
ci_interval.}

\item{adj2}{Value controlling the total width of the plot.}

\item{order}{String similar to the one used in metapipeline::mining_
diamonds() order.}

\item{...}{Other arguments for the geom_polygon().}
}
\value{
An object of class ggplot.
}
\description{
Forest plot showing the summary results of
the dataframe obtained from rma_list_to_summary_df().
}
\examples{
#' library(metafor)
order <- c("DFS", "OS")
diamonds <- metaPipeline::auto_rma_uni(dat.riley2003, group = "outcome") |>
   metaPipeline::rma_list_to_summary_df() |>
   metaPipeline::mining_diamonds(main_group = "outcome", order = order) |>
   metaPipeline::plot_diamonds(
     title = "dat.riley2003 from metafor package",
     show_grids_v = TRUE, xlab = "SMD", group1 = group_plot,
     order = order, vline_ref = 0, adj1 = 1.2, adj2 = 16, ci_interval = TRUE)
}
