% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{post process for FE or REML Random Effect meta analysis}
\usage{
post_process(result, combined_data, model, significance_level = 0.05,
  plot_forest = TRUE, output_path = "./", phenotype = NULL,
  gender = NULL, print_log = FALSE, log_path = "./",
  verbose = TRUE)
}
\arguments{
\item{result}{(data.frame) result from \code{\link{meta}} using FE and including FDR}

\item{combined_data}{(data.frame) data used to create the above result from \code{\link{load_files}}}

\item{model}{(string) name of the model (can be either FE or REML)}

\item{significance_level}{(double) FDR significance level for cut off}

\item{plot_forest}{(bool) TRUE: Draws a forest plot of the top hits FALSE: won't}

\item{output_path}{(string) output path for the forest plot}

\item{phenotype}{(string) phenotype currently examined (for the forest plot name and label)}

\item{gender}{(string) gender currently examined (for the forest plot and label)}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: Prints output to terminal FALSE: won't}
}
\description{
post processing function for \code{FE} and \code{REML} model which draws a forest plot for all site whose FDR adjusted p-value is above \code{significance_level}
}
\author{
Antoine Weihs <uni.antoine.weihs@gmail.com>
}
