% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_Z.R
\name{post_processZ}
\alias{post_processZ}
\title{post process for Z-based meta analysis}
\usage{
post_processZ(
  result,
  FDR,
  significance_level = 0.05,
  plot_manhattan = TRUE,
  plot_annotation = TRUE,
  return_hits_only = FALSE,
  output_path = "./",
  annotate_result = TRUE,
  phenotype = NULL,
  stratum = NULL,
  print_log = FALSE,
  log_path = "./",
  verbose = TRUE
)
}
\arguments{
\item{result}{(data.frame) result from \code{\link{meta}} using Z}

\item{FDR}{(bool) TURE: FDR values are used instead of p-values (input needs to contain an FDR column) FALSE: p-values are used}

\item{significance_level}{(double) FDR/p-value significance level for cut off}

\item{plot_manhattan}{(bool) TRUE: will create a manhattan plot of the results FALSE: won't}

\item{plot_annotation}{(bool) TRUE: creates annotaton plots of top hits FALSE: won't}

\item{return_hits_only}{(bool) TRUE: will only return sites above the significance level cut off. FALSE: will return the whole data set.}

\item{output_path}{(string) output path for the forest plot}

\item{annotate_result}{(bool) TRUE: will merge output file with annoation file FALSE: won't}

\item{phenotype}{(string) phenotype currently examined (for the forest plot name and label)}

\item{stratum}{(string) stratum currently examined (for the forest plot and label)}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: Prints output to terminal FALSE: won't}
}
\value{
data.frame containing the original \code{result} data set if \code{annotate_result} = FALSE or the \code{result} data.frame merged with the annotation data.frame if
        \code{annotate_result} = TRUE.
}
\description{
post processing function for \code{Z} model which can draw
             draws a manhattan plot and/or annotates the output with an annotation file
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
