% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preMeta_QC.R
\name{preMeta_QC}
\alias{preMeta_QC}
\title{Pre meta-analysis QC pipeline}
\usage{

  preMeta_QC(data_summary_path = "/home/weihsa/data/Thyroid_Cohorts/Cohort_Summary/summary.txt",
  save_path = "./", phenotype, stratum, cohort = NULL,
  anno_file_path = "/home/weihsa/data/EWAS/Ilmn_Methylation_Masterfile-180309.txt.gz",
  pre_filter_step = FALSE, pre_filter_script, run_filter = TRUE,
  filter_percentage = 25, pre_plotting_step = FALSE,
  pre_plotting_script, plotting = TRUE, plotting_chr = TRUE,
  same_scale = TRUE, plot_trim = FALSE, pre_printing_step = FALSE,
  pre_printing_script, print_summary = TRUE, print_flags = TRUE,
  flags_num_NA = 1000, flags_min_beta = -2, flags_max_beta = 2,
  flags_max_se = 4, flags_min_inflation = 0.9,
  flags_max_inflation = 1.1, flags_decimal_places = 4,
  final_step = FALSE, final_script, print_log = FALSE,
  log_path = "./log.txt", verbose = TRUE)
}
\arguments{
\item{data_summary_path}{(string) path to the file that contains all the info of the cohorts. Has to include the following tab separated
columns: \itemize{
\item{\code{cohort}}{: Name of the cohort}
\item{\code{phenotype}}{: Name of the observed phenotype used in the file}
\item{\code{stratum}}{ Name of the stratum used in the file}
\item{\code{array_type}}{: type of array used by the cohort. Can be 450 or EPIC}
\item{\code{probeID}}{: Name of the column containing the probeIDs}
\item{\code{beta}}{: Name of the column containing the effect sizes}
\item{\code{se}}{: Name of the column containing the standard error values}
\item{\code{p_val}}{: Name of the column containing the p-values}
\item{\code{size}}{: Name of the column containing the sample sizes}
\item{\code{separator}}{: type of separator used in the file. Example: ","}
\item{\code{filename}}{: Full name of the file. Example: "example_name.csv"}
\item{\code{file_path}}{: Full path to the file. Example: "/home/user/file_location/"}
}}

\item{save_path}{(string) place where the outputs are saved}

\item{phenotype}{(string) phenotype that should be analysed (has to appear in the summary$Phenotype column)}

\item{stratum}{(string) stratum that should be analysed (has to appear in the summary$stratum column)}

\item{cohort}{(string) if only certain cohorts of the summary file should be analysed (has to appear in the summary$cohort column)}

\item{anno_file_path}{(string) path to the annotation file}

\item{pre_filter_step}{(bool) TRUE: runs script located at \code{pre_filter_script} FALSE: won't}

\item{pre_filter_script}{(string) path to script. \code{pre_filter_step} has to be TRUE to run}

\item{run_filter}{(bool) TRUE: runs size filter, removing sites with a sample size of less then \code{filter_percentage} of the maximum sample size
FALSE: won't}

\item{filter_percentage}{(int) percentage used to filter}

\item{pre_plotting_step}{(bool) TRUE: runs script located at \code{pre_plotting_script} FALSE: won't}

\item{pre_plotting_script}{(string) path to script. \code{pre_plotting_step} has to be TRUE to run}

\item{plotting}{(bool) TRUE: plots will be generated FALSE: plots won't be generated}

\item{plotting_chr}{(bool) TRUE: chromosome plots will be generated FALSE: won't}

\item{same_scale}{(bool) TRUE: PVAL, BETA and SE separate plots are plotted using the same scale on the axes. FALSE: the separate plots are drawn on
different scale axes}

\item{plot_trim}{(bool) TRUE: will trim the density plots at local min and max FALSE: will plot density curve from global min to global max}

\item{pre_printing_step}{(bool) TRUE: runs script located at \code{pre_printing_script} FALSE: won't}

\item{pre_printing_script}{(string) path to script. \code{pre_printing_step} has to be TRUE to run}

\item{print_summary}{(bool) TRUE: file containing cohort summaries will be generated False: won't}

\item{print_flags}{(bool) TRUE: file containing flags will be generated FALSE: won't}

\item{flags_num_NA}{(int) parameter for \code{print_flags()}}

\item{flags_min_beta}{(int) parameter for \code{print_flags()}}

\item{flags_max_beta}{(int) parameter for \code{print_flags()}}

\item{flags_max_se}{(int) parameter for \code{print_flags()}}

\item{flags_min_inflation}{(int) parameter for \code{print_flags()}}

\item{flags_max_inflation}{(int) parameter for \code{print_flags()}}

\item{flags_decimal_places}{(int) parameter for \code{print_flags()}}

\item{final_step}{(bool) TRUE: runs script located at \code{final_script} FALSE: won't}

\item{final_script}{(string) path to script. \code{final_step} has to be TRUE to run}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: print log to command line FALSE: won't}
}
\description{
This function combines multiple functions into one to perform an pre meta-analysis quality control. The script is written in a modular
             fashion and multiple entry points have been added to expand the script where needed. Please note that before adding new scripts at an
             entry point, it is advisable to check the source code it is advisable to check the source code to make sure the script is compatible. The
             steps of the script are: load_files -> pre_filter entry point -> run_filter -> pre_plotting entry point -> plotting -> plotting chromosomes
             -> pre_printing entry point -> printing summary -> printing flags -> final entry point
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
