% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print_flags}
\alias{print_flags}
\title{Print flags}
\usage{
print_flags(data_set, num_NA = 1000, min_beta = -2, max_beta = 2,
  max_se = 4, min_inflation = 0.8, max_inflation = 1.2,
  decimal_places = 4, save_path, gender, phenotype, verbose = TRUE,
  print_log = FALSE, log_path = "./log.txt")
}
\arguments{
\item{data_set}{(data.frame) data set created by \code{\link{load_files}}}

\item{num_NA}{(int) threshold for the number of NAs}

\item{min_beta}{(int) lower threshold for BETA values}

\item{max_beta}{(int) upper threshold for BETA values}

\item{max_se}{(int) upper threshold for SE values}

\item{min_inflation}{(int) lower threshold values for \code{\link[bacon]{bacon}} inflation}

\item{max_inflation}{(int) upper threshold values for \code{\link[bacon]{bacon}} inflation}

\item{decimal_places}{(int) lower threshold for average number of decimal places}

\item{save_path}{(string) place where the outputs are saved}

\item{gender}{(string) gender currently analysed}

\item{phenotype}{(string) phenotype currently analysed}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log}
}
\description{
This function prints flags to an output file. Used for the pre-meta quality control analysis.
              At the moment these flags are:
              \itemize{
              \item{Different number of NAs in BETA, Pval and SE}
              \item{Number of NAs above \code{num_NA}}
              \item{min(BETA) below \code{min_beta} or max(BETA) above \code{max_beta}}
              \item{max(SE) above \code{max_se}}
              \item{more then 1 SE or BETA value above 75th + 1.5*IQR or below 25th percentile - 1.5*IQR}
              \item{\code{\link[bacon]{bacon}} inflation below \code{min_inflation} or above \code{max_inflation}}
              \item{number of cites of one chromosome = 0}
              \item{negative standard errors or p values}
              \item{average number of decimal places of BETA, Pval and SE smaller then \code{decimal_places}}
              }
}
\author{
Antoine Weihs <uni.antoine.weihs@gmail.com>
}
