% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print_summaries}
\alias{print_summaries}
\title{Print Summaries}
\usage{
print_summaries(data_set, save_path, stratum, phenotype, verbose = TRUE,
  print_log = FALSE, log_path = "./log.txt")
}
\arguments{
\item{data_set}{(data.frame) data set created by \code{\link{load_files}}}

\item{save_path}{(string) place where the outputs are saved}

\item{stratum}{(string) stratum currently analysed}

\item{phenotype}{(string) phenotype currently analysed}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log}
}
\description{
This function prints summary statistics to an output file. Used for the pre-meta quality control analysis.
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
