% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pval_plot}
\alias{pval_plot}
\title{Pvalue plots}
\usage{
pval_plot(
  data_set,
  same_scale = TRUE,
  save_path,
  stratum,
  phenotype,
  trim = FALSE,
  verbose = TRUE,
  print_log = FALSE,
  log_path = "./log.txt"
)
}
\arguments{
\item{data_set}{(data.frame) data set created by \code{\link{load_files}}}

\item{same_scale}{(bool) True: the separate plots are all drawn using the same axis scales False: the separate plots are all drawn using different scales}

\item{save_path}{(string) place where the outputs are saved}

\item{stratum}{(string) stratum currently analysed}

\item{phenotype}{(string) phenotype currently analysed}

\item{trim}{(bool) TRUE: will trim the density plots at local min and max FALSE: will plot density curve from global min to global max}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log}
}
\description{
This function creates plots of the p-value distributions. Used for the pre-meta quality control analysis.
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
