% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma_list_to_summary_df.R
\name{rma_list_to_summary_df}
\alias{rma_list_to_summary_df}
\title{rma_list_to_summary_df()}
\usage{
rma_list_to_summary_df(rma_list)
}
\arguments{
\item{{rma_list}}{A list containing rma.uni results from
metaPipeline::auto_rma_uni() with groups.}
}
\value{
A dataframe containing a summary of rma results.
}
\description{
Summarize the list of rma results.
}
