% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bacon.R
\name{run_bacon}
\alias{run_bacon}
\title{Apply BACON}
\usage{
run_bacon(data_set, run = TRUE, print_log = FALSE,
  log_path = "log.txt", verbose = TRUE)
}
\arguments{
\item{data_set}{(data.frame) data set created by \code{\link{load_files}}}

\item{run}{(bool) TRUE: this function is applied, FALSE: this function won't be applied}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: print log to command line FALSE: won't}
}
\value{
\code{data_set} with BETA and SE values modified if BACON-inflation is above 1. NOTE: Sites where effect or standard error was NA have been removed
}
\description{
This function uses the \code{\link[bacon]{bacon}} package to calculate the inflation and bias and modifies the input data if inflation is above 1
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
