% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{run_filter}
\alias{run_filter}
\title{Filter data}
\usage{
run_filter(data_set, percentage = 25, run = TRUE, print_log = FALSE,
  log_path = "./log.txt", verbose = TRUE)
}
\arguments{
\item{data_set}{(data.frame) data set created by \code{\link{load_files}}}

\item{percentage}{(int) percentage of the maximal sample size that has to be present in the sample site of each site}

\item{run}{(bool) TRUE: this function is applied, FALSE: this function won't be applied}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: print log to command line FALSE: won't}
}
\value{
filtered \code{data_set} NOTE: Sites with NA in effect or standard error column have been removed
}
\description{
This function filters the data, removing all sites that have a sample size below the threshold value
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
