% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_meta.R
\name{run_meta}
\alias{run_meta}
\title{Meta-analysis pipeline}
\usage{
run_meta(phenotype, stratum, num_cores = 1, model = "FE",
  data_summary_path = "/home/weihsa/data/Thyroid_Cohorts/Cohort_Summary/summary.txt",
  cohort = NULL, cpg_filter = NULL, pre_bacon_step = FALSE,
  pre_bacon_path, apply_bacon = TRUE, pre_filter_step = FALSE,
  pre_filter_path, filter = TRUE, filter_percentage = 25, FDR = TRUE,
  pre_meta_step = FALSE, pre_meta_path, post_meta_step = FALSE,
  post_meta_path, run_post_process = FALSE, run_posterior_Check = TRUE,
  post_sigLevel = 0.05, post_cutoff = 0.95, post_replicates = 1000,
  plot_forest = FALSE, plot_manhattan = TRUE,
  annotate_result = FALSE, annotation_filepath = NULL,
  return_hits_only = FALSE, final_step = FALSE, final_path,
  print_log = TRUE, save_RData = FALSE, save_csvData = FALSE,
  output_path = "./", verbose = TRUE)
}
\arguments{
\item{phenotype}{(string) phenotype that should be analysed (has to appear in the summary$Phenotype column)}

\item{stratum}{(string) stratum that should be analysed (has to appear in the summary$stratum column)}

\item{num_cores}{(int) number of cores that should be used for the meta analysis (only works on Linux)}

\item{model}{(string) Estimator model for tau. Standard \code{FE}. Otherwise \code{BAYES} for a Bayesian approach,
\code{REML} for a restricted maximum likelihood approach (not recommended for small number of cohorts (<10)) or \code{Z} for p-value based approach}

\item{data_summary_path}{(string) path to the file that contains all the info of the cohorts. Has to include the following tab separated
columns: \itemize{
\item{\code{cohort}}{: Name of the cohort}
\item{\code{phenotype}}{: Name of the observed phenotype used in the file}
\item{\code{stratum}}{: Name of the stratum used in the file}
\item{\code{array_type}}{: type of array used by the cohort. Can be 450 or EPIC}
\item{\code{probeID}}{: Name of the column containing the probeIDs}
\item{\code{beta}}{: Name of the column containing the effect sizes}
\item{\code{se}}{: Name of the column containing the standard error values}
\item{\code{p_val}}{: Name of the column containing the p-values}
\item{\code{size}}{: Name of the column containing the sample sizes}
\item{\code{separator}}{: type of separator used in the file. Example: ","}
\item{\code{filename}}{: Full name of the file. Example: "example_name.csv"}
\item{\code{file_path}}{: Full path to the file. Example: "/home/user/file_location/"}
}}

\item{cohort}{(string vector) if only certain cohorts of the summary file should be analysed (has to appear in the summary$Cohort column)}

\item{cpg_filter}{(string vector) if only a certain CpG site shall be analysed}

\item{pre_bacon_step}{(bool) TRUE: runs script located at \code{pre_bacon_path} FALSE: won't}

\item{pre_bacon_path}{(string) path to script. \code{pre_bacon_step} has to be TRUE to run}

\item{apply_bacon}{(bool) PREPROCESS: TRUE: \code{\link[bacon]{bacon}} will be run on cohorts that meet certain criteria FALSE: \code{\link[bacon]{bacon}} won't be applied}

\item{pre_filter_step}{(bool) TRUE: runs script located at \code{pre_filter_path} FALSE: won't}

\item{pre_filter_path}{(string) path to script. \code{pre_filter_step} has to be TRUE to run}

\item{filter}{(bool) PREPROCESS: TRUE: \code{\link{run_filter}} will be applied FALSE: won't}

\item{filter_percentage}{(int) PREPROCESS: if \code{filter} is TRUE -> percentage of the maximal sample size that has to be present in the
sample site of each site}

\item{FDR}{(bool) TRUE: p values will be corrected for FDR, FALSE: won't}

\item{pre_meta_step}{(bool) TRUE: runs script located at \code{pre_meta_path} FALSE: won't}

\item{pre_meta_path}{(string) path to script. \code{post_meta_path} has to be TRUE to run}

\item{post_meta_step}{(bool) TRUE: runs script located at \code{pre_meta_path} FALSE: won't}

\item{post_meta_path}{(string) path to script. \code{post_meta_step} has to be TRUE to run}

\item{run_post_process}{(bool) TRUE: applies post_process function, FALSE: Won't
NOTE: Significant runtime increase if applied using the BAYES model and \code{run_posterior_Check} = TRUE}

\item{run_posterior_Check}{(bool) TRUE: will run a posterior predictive (see \link[bayesmeta]{pppvalue}) (NOTE: computationally intensive) FALSE: won't}

\item{post_sigLevel}{(double) POSTPROCESS: if \code{run_post_process} is TRUE and \code{model} is FE or REML-> FDR significance level for cut off}

\item{post_cutoff}{(double) POSTPROCESS: if \code{run_post_process} is TRUE and \code{model} is BAYES -> effect posterior probability cutoff}

\item{post_replicates}{(int) POSTPROCESS: if \code{run_post_process} is TRUE and \code{model} is BAYES -> number of replicates performed during the
posterior probability check (n >> 100 recommended)}

\item{plot_forest}{(bool) POSTPROCESS: TRUE: will create forest plot saved to output_path if run_post_process is TRUE FALSE: won't}

\item{plot_manhattan}{(bool) POSTPROCESS: TRUE: will create a manhattan plot of the results if model is FE or REML FALSE: won't}

\item{annotate_result}{(bool) POSTPROCESS: TRUE: will merge output file with annoation file FALSE: won't}

\item{annotation_filepath}{(string) POSTPROCESS: file path to annotation file (annotation file has to contain a 'Markername' column, has to be a tab separated file,
comments have to be preceded by '#')}

\item{return_hits_only}{(bool) POSTPROCESS: TRUE: If \code{run_post_process} is TRUE will only return sites above the significance level cut off. FALSE: will return
the whole data set.}

\item{final_step}{(bool) TRUE: runs script located at \code{final_path} FALSE: won't}

\item{final_path}{(string) path to script. \code{final_step} has to be TRUE to run}

\item{print_log}{(bool) TRUE: log will be printed to output_path FALSE: no log will be printed}

\item{save_RData}{(bool) TRUE: result will be saved as .RData file to output_path, FALSE: result won't be saved}

\item{save_csvData}{(bool) TRUE: result will be saved as .csv file to output_path, FALSE: result won't be saved}

\item{output_path}{(string) path were the result should be saved}

\item{verbose}{(bool) TRUE: will print information to the terminal FALSE: won't}
}
\value{
data.frame containing the results of the pipeline (see \code{\link{meta}} and \code{\link{bayes_post_process}}
}
\description{
Main script that handles the meta analysis. The script is written in a modular fashion and multiple entry points have been added to
             expand the script where needed. Please note that before adding new scripts at an entry point, it is advisable to check the source code
             to make sure the script is compatible. The steps of the script are: load files -> pre-filter entry point -> run filter -> pre-bacon
             entry point -> run bacon -> pre-meta-analysis entry point -> run meta-analysis -> post-meta-analysis entry point -> post process
             -> final entry point -> save results
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
