% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_files.R
\name{save_files}
\alias{save_files}
\title{Save file functions}
\usage{
save_files(data, type, save_path, phenotype, gender, print_log, log_path,
  verbose)
}
\arguments{
\item{data}{(any) data set to be saved}

\item{type}{(string) either \code{csv} or \code{rdata}. File type \code{data} should be saved as}

\item{save_path}{(sting) full path to which the output should be saved to (including filename)}

\item{phenotype}{(string) phenotype used in the analysis}

\item{gender}{(string) gender used in the analysis}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: Prints output to terminal FALSE: won't}
}
\description{
functions that will save files as *.RData or *.csv
}
\author{
Antoine Weihs <uni.antoine.weihs@gmail.com>
}
