% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{post process for FE or REML Random Effect meta analysis}
\usage{
post_process(result, combined_data, model, FDR,
  significance_level = 0.05, plot_forest = TRUE,
  plot_manhattan = TRUE, output_path = "./", annotate_result = TRUE,
  annotation_filepath = NULL, return_hits_only = FALSE,
  phenotype = NULL, stratum = NULL, print_log = FALSE,
  log_path = "./", verbose = TRUE)
}
\arguments{
\item{result}{(data.frame) result from \code{\link{meta}} using FE or REML}

\item{combined_data}{(data.frame) data used to create the above result from \code{\link{load_files}}}

\item{model}{(string) name of the model (can be either FE or REML)}

\item{FDR}{(bool) TRUE: FDR values are used instead of p-values (input needs to contain an FDR column) FALSE: p-values are used}

\item{significance_level}{(double) FDR/p-value significance level for cut off}

\item{plot_forest}{(bool) TRUE: Draws a forest plot of the top hits FALSE: won't}

\item{plot_manhattan}{(bool) TRUE: will create a manhattan plot of the results FALSE: won't}

\item{output_path}{(string) output path for the forest plot}

\item{annotate_result}{(bool) TRUE: will merge output file with annoation file FALSE: won't}

\item{annotation_filepath}{(string) file path to annotation file (annotation file has to contain a 'Markername' column, has to be a tab separated file,
comments have to be preceded by '#')}

\item{return_hits_only}{(bool) TRUE: will only return sites above the significance level cut off. FALSE: will return the whole data set.}

\item{phenotype}{(string) phenotype currently examined (for the forest plot name and label)}

\item{stratum}{(string) stratum currently examined (for the forest plot and label)}

\item{print_log}{(bool) TRUE: print to log file FALSE: won't}

\item{log_path}{(string) path to log file}

\item{verbose}{(bool) TRUE: Prints output to terminal FALSE: won't}
}
\value{
data.frame containing the original \code{result} data set if \code{annotate_result} = FALSE or the \code{result} data.frame merged with the annotation data.frame if
        \code{annotate_result} = TRUE.
}
\description{
post processing function for \code{FE} and \code{REML} model which can draw
             a forest plot for all site whose FDR-value or P-value is above \code{significance_level} and/or
             draws a double manhattan plot and/or annotates the output with an annotation file
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
