% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{se_vs_size_plot}
\alias{se_vs_size_plot}
\title{SE vs Size plots}
\usage{
se_vs_size_plot(data_set, save_path, stratum, phenotype, verbose = TRUE,
  print_log = FALSE, log_path = "./log.txt")
}
\arguments{
\item{data_set}{(data.frame) data set created by \code{\link{load_files}}}

\item{save_path}{(string) place where the outputs are saved}

\item{stratum}{(string) stratum currently analysed}

\item{phenotype}{(string) phenotype currently analysed}

\item{verbose}{(bool) TRUE: will print output to terminal FALSE: won't}

\item{print_log}{(bool) TRUE: prints logs to log_path FALSE: won't}

\item{log_path}{(string) full path + file name of log}
}
\description{
This function creates plots the median cohort size versus the median cohort standard error. Used for the pre-meta quality control analysis.
}
\author{
Antoine Weihs <antoine.weihs@uni-greifswald.de>
}
